requirements.txt
aiogram==3.3.0
aiohttp
yookassa==3.0.1
python-dotenv
